+++ src/main/java/lk/pos/model/Product.java
@@ -7,9 +7,9 @@ public class Product {
     private final IntegerProperty id = new SimpleIntegerProperty();
     private final StringProperty name = new SimpleStringProperty();
     private final DoubleProperty price = new SimpleDoubleProperty();
-    private final IntegerProperty qty = new SimpleIntegerProperty();
+    private final DoubleProperty qty = new SimpleDoubleProperty();
 
-    public Product(int id, String name, double price, int qty) {
+    public Product(int id, String name, double price, double qty) {
         this.id.set(id);
         this.name.set(name);
         this.price.set(price);
@@ -29,7 +29,7 @@ public class Product {
         return price;
     }
 
-    public IntegerProperty qtyProperty() {
+    public DoubleProperty qtyProperty() {
         return qty;
     }
 
@@ -37,5 +37,5 @@ public class Product {
     public int getId() { return id.get(); }
     public String getName() { return name.get(); }
     public double getPrice() { return price.get(); }
-    public int getQty() { return qty.get(); }
+    public double getQty() { return qty.get(); }
 }
