+++ src/main/java/lk/pos/util/DatabaseInitializer.java
@@ -21,7 +21,7 @@ public class DatabaseInitializer {
                 id INT AUTO_INCREMENT PRIMARY KEY,
                 name VARCHAR(100) NOT NULL,
                 price DECIMAL(10,2) NOT NULL,
-                qty INT NOT NULL
+                qty DECIMAL(10,3) NOT NULL
             )
         """;
